lib/vfscore: Use mount definitions from <sys/mount.h>
Give mount constant/flags definitions introduced by the libC higher
priority than the internal ones. This is done to provide a correct
libC API implementation with vfscore.

+++ lib/vfscore/mount.c
@@ -109,7 +109,8 @@ int device_close(struct device *dev)
 }
 
 int
-mount(const char *dev, const char *dir, const char *fsname, int flags, const void *data)
+mount(const char *dev, const char *dir, const char *fsname, unsigned long flags,
+      const void *data)
 {
 	const struct vfscore_fs_type *fs;
 	struct mount *mp;
